pokedex/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
gunicorn==20.1.0
idna==3.3
importlib-resources==5.7.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
pytest-django==4.5.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
